RevB - Added front window cutout for the pulleys
I wanted to be able to see how the belt was tracking on the pulleys, but still retain the ability to self feed the belts. So I added a tool that plugs into the window when feeding the belts.

Also added a set of +0.2mm files to compensate for parts that print short.